refactor(trusty-mpm): split tm.rs monolith (4,638 lines) into focused submodules (closes #395)
Convert src/bin/tm.rs into src/bin/tm/ module directory with 17 focused
files, every one under the 500-line cap. Pure mechanical refactor — zero
behavior changes.

Layout after split:
  src/bin/tm/main.rs          133  thin entry point + dispatch
  src/bin/tm/cli.rs           446  all clap Parser/Subcommand definitions
  src/bin/tm/types.rs          47  SessionRow, ProjectRow, EventRow DTOs
  src/bin/tm/formatters/
    mod.rs                     10
    session.rs                 72  short_id, event_summary, compression stats
    services.rs               112  print_services_table, print_service_status_block
    banner.rs                 311  banner rendering, terminal_width
  src/bin/tm/commands/
    mod.rs                     18
    daemon.rs                ~390  start/stop/restart/status, PID helpers
    daemon_run.rs            ~193  run_daemon and private boot helpers (split from daemon.rs)
    install.rs               ~320  install, hook additions, deploy report
    launch.rs                ~343  launch, connect
    misc.rs                  ~360  status, events, doctor, hook, coordinator, etc.
    project.rs               ~132  project subcommand
    services.rs              ~201  services subcommand
    session.rs               ~434  session subcommand, compose_session_instructions
    telegram.rs              ~111  telegram subcommand
  tests split into three files (each <500 lines):
    tests.rs                ~505  CLI parse + formatter unit tests
    tests_behavior_a.rs     ~450  install / hook / session output tests
    tests_behavior_b.rs     ~295  attach / connect / daemon / services / #382

Also:
- Update [[bin]] path in Cargo.toml: src/bin/tm.rs → src/bin/tm/main.rs
- Remove crates/trusty-mpm/src/bin/tm.rs from .line-cap-allowlist.tsv (ratchet down)

Quality bar: cargo check + clippy (0 warnings) + fmt --check + all 958 tests pass
+ check_line_cap.sh (0 violations).
